# packages/api-client/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fullstack Starter API",
    "description": "API documentation for Fullstack Starter Backend",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/billing/checkout": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "session_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "complete"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "pending"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "failed"
                              ]
                            }
                          ]
                        }
                      },
                      "required": [
                        "status"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response format",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Error response format",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "description": "Human-readable error message",
                      "type": "string",
                      "example": "User already exists"
                    },
                    "code": {
                      "description": "Machine-readable error code for programmatic handling",
                      "type": "string",
                      "example": "USER_EXISTS"
                    },
                    "details": {
                      "description": "Additional error context (structure varies by error type)",
                      "type": "object",
                      "properties": {
                        "errors": {
                          "description": "Array of validation errors if present",
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "description": "Field name that failed validation",
                                "type": "string",
                                "example": "email"
                              },
                              "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Invalid format"
                              }
                            },
                            "required": [
                              "field",
                              "message"
                            ]
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "priceId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response format",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Error response format",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "description": "Human-readable error message",
                      "type": "string",
                      "example": "User already exists"
                    },
                    "code": {
                      "description": "Machine-readable error code for programmatic handling",
                      "type": "string",
                      "example": "USER_EXISTS"
                    },
                    "details": {
                      "description": "Additional error context (structure varies by error type)",
                      "type": "object",
                      "properties": {
                        "errors": {
                          "description": "Array of validation errors if present",
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "description": "Field name that failed validation",
                                "type": "string",
                                "example": "email"
                              },
                              "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Invalid format"
                              }
                            },
                            "required": [
                              "field",
                              "message"
                            ]
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}